[BSE-5] zero-copy DATE32 support for arrow (#5253)
Changes the default representation of dates to be 32-bit signed integers representing the offset in days since the day of the Unix epoch (1970-01-01). This matches the in-memory Arrow representation so we can share buffers and avoid copies during IO.

* [BE-4515] [BE-4519] Use arrow format for dates internally

* [BE-4517] Modify non-serialization C++ code to handle 32-bit dates (#5208)

* [BE-4518] Date32 support in Arrow (#5227)

* Switch to new array allocation type

* Move min_row_number_filter with multiple order by columns to use 32 bit date

updated row_number_filter code

---------

+++ bodo/libs/_bodo_common.cpp
@@ -26,7 +26,7 @@ void bodo_common_init() {
     numpy_item_size[Bodo_CTypes::FLOAT64] = sizeof(double);
     numpy_item_size[Bodo_CTypes::DECIMAL] = BYTES_PER_DECIMAL;
     numpy_item_size[Bodo_CTypes::DATETIME] = sizeof(int64_t);
-    numpy_item_size[Bodo_CTypes::DATE] = sizeof(int64_t);
+    numpy_item_size[Bodo_CTypes::DATE] = sizeof(int32_t);
     // TODO: [BE-4106] TIME size should depend on precision.
     numpy_item_size[Bodo_CTypes::TIME] = sizeof(int64_t);
     numpy_item_size[Bodo_CTypes::TIMEDELTA] = sizeof(int64_t);
